Crate redirect

Modules

command
GPU command encapsulation.

descriptor
a relatively small block of data that fully describles an object to the GPU.

device
defines Device, interface for a 3D display adapter

error
error types

factory
this module contains Factory, the API entry point, along with several other fundamental structures such as Adapter and Output.

fence
a fence is used for synchronization of CPUs and GPUs

format
formats used by direct3d

pipeline
describes graphics and compute pipelines

resource
resource

shader
executable code snippet for GPU

swapchain
link between the graphics API and the target surface