Module redirect::command

GPU command encapsulation.

Structs

Bundle
a command list bundle

BundleCommandAllocator
An allocator for GPU commands

BundleRecording
a command list bundle during recording state

BundleRecordingWithHeap
a command list bundle during recording state with descriptor heap already set

CommandListType
type of a CommandList

CommandQueue
A GPU command queue, providing methods for command submission, execution synchronization, etc.

CommandQueueDesc
describes a command queue.

CommandQueueFlags
misc flags for a command queue

CommandQueuePriority
priority levels for a command queue

DepthStencilClearFlags
depth stencil clear flags

DirectCommandAllocator
An allocator for GPU commands

DirectCommandList
a direct command list

DirectCommandListRecording
a direct command list on recording state

Constants

COMMAND_LIST_TYPE_BUNDLE
bundle inherits all GPU state (except PSO and primitive topology) from the direct list, where it must be opearated on

COMMAND_LIST_TYPE_COMPUTE
computing command list

COMMAND_LIST_TYPE_COPY
copying(drawing) command list

COMMAND_LIST_TYPE_DIRECT
direct list doesn't inherit any GPU state.

COMMAND_QUEUE_FLAG_DISABLE_GPU_TIMEOUT

COMMAND_QUEUE_FLAG_NONE

COMMAND_QUEUE_PRIOIRITY_GLOBAL_REALTIME

COMMAND_QUEUE_PRIOIRITY_NORMAL

COMMAND_QUEUE_PRIORITY_HIGH

DS_CLEAR_FLAG_DEPTH

DS_CLEAR_FLAG_STENCIL

Traits

CommandList

CommandListWithHeap
common methods for a command list with bounded descriptor heaps

ComputeCommandList
common methods for a compute command list

CopyCommandList
common methods for a copy command list

GraphicsCommandList
common methods for a graphics command list