Struct redirect::command::DirectCommandListRecording

pub struct DirectCommandListRecording<'a> {
    pub ptr: ComPtr<ID3D12GraphicsCommandList>,
    pub alloc: &'a mut DirectCommandAllocator,
    pub initial_state: Option<&'a PipelineState>,
}

a direct command list on recording state

Fields
ptr: ComPtr<ID3D12GraphicsCommandList> alloc: &'a mut DirectCommandAllocator

command allocator used to back up command recording

initial_state: Option<&'a PipelineState>

initial state of this command list

Methods

impl<'a> DirectCommandListRecording<'a>

fn clear_dsv(
&mut self,
dsv: CpuDsvHandle,
flags: DepthStencilClearFlags,
depth: f32,
stencil: u8,
rects: Option<&[Rect]>
)

record a command to clear the dsv. more info

fn clear_uav_f32(
&mut self,
heap: &mut CbvSrvUavHeap,
index: u32,
resource: &mut RawResource,
values: &[f32; 4],
rects: Option<&[Rect]>
)

record a command to clear uav. more info

fn clear_uav_u32(
&mut self,
heap: &mut CbvSrvUavHeap,
index: u32,
resource: &mut RawResource,
values: &[u32; 4],
rects: Option<&[Rect]>
)

record a command to clear uav. more info

fn clear_rtv(
&mut self,
rtv: CpuRtvHandle,
values: &[f32; 4],
rects: Option<&[Rect]>
)

record clearing a rtv

fn copy_resource(&mut self, dst: &mut RawResource, src: &mut RawResource)

record a resource copy operation.

restrictions

must be different resources

must be the same type

must have identical dimensions

must have compatible formats

more

fn copy_buffer_region(
&mut self,
dst: &mut RawResource,
dst_offset: u64,
src: &mut RawResource,
src_offset: u64,
size: u64
)

record a buffer copy operation

fn copy_texture_region(
&mut self,
dst: TextureCopyLocation,
dstx: u32,
dsty: u32,
dstz: u32,
src: TextureCopyLocation,
src_box: &Box3u
)

record a texture copy operation. more info

fn execute_bundle(&mut self, bundle: &Bundle)

execute a bundle. more

fn so_set_targets(&mut self, start_slot: u32, sovs: &[StreamOutputBufferView])

set the stream output buffer views

fn rs_set_scissors(&mut self, scissors: &[Rect])

set scissor rectangles

fn rs_set_viewports(&mut self, viewports: &[Viewport])

set viewports

fn resolve_ms(
&mut self,
dst: &mut RawResource,
dst_sub: u32,
src: &mut RawResource,
src_sub: u32,
format: DxgiFormat
)

resolve a multisampled resource into a non-MS resource. more

fn resource_barriers(&mut self, barriers: &ResourceBarriersBuilder)

synchronizaing multiple access to resources. more

fn clear_state(&mut self)

reset the state of a direct command list back to the state it was in when created. initial_state has to match this state

fn set_descriptor_heaps(
&mut self,
cbv_srv_uav_heap: Option<&CbvSrvUavHeap>,
rtv_heap: Option<&RtvHeap>,
dsv_heap: Option<&DsvHeap>,
sampler_heap: Option<&SamplerHeap>
)

fn reset<'b>(
self,
alloc: &'b mut DirectCommandAllocator,
initial_state: Option<&'b PipelineState>
) -> Result<DirectCommandListRecording<'b>, (WinError, Self)>

reset a command list back to the initial state. more

fn close(self) -> Result<DirectCommandList, WinError>

close the current recording