Struct redirect::command::GraphicsCommandList

pub struct GraphicsCommandList {
    pub ptr: ComPtr<ID3D12GraphicsCommandList>,
}

a list of graphics commands, including execution intrumenting APIs etc. TODO: distinct bundle from direct command lists at type level?

Fields
ptr: ComPtr<ID3D12GraphicsCommandList>
Methods

impl GraphicsCommandList

fn set_graphics_rootsig(&mut self, rootsig: &RootSig)

set the graphics root signature. TODO: distince root signatures

fn set_graphics_root_constant(
&mut self,
param_index: u32,
value: u32,
param_offset: u32
)

set a constant in the graphics root signature

fn set_graphics_root_constants(
&mut self,
param_index: u32,
values: &[u32],
param_offset: u32
)

set a group of constants in the graphics root signature

fn set_graphics_root_cbv(
&mut self,
param_index: u32,
resource: &mut RawResource
)

set a cbv in the graphics root signature

fn set_graphics_root_srv(
&mut self,
param_index: u32,
resource: &mut RawResource
)

set a srv in the graphics root signature

fn set_graphics_root_uav(
&mut self,
param_index: u32,
resource: &mut RawResource
)

set a uav in the graphics root signature

fn set_graphics_root_dt(
&mut self,
param_index: u32,
base_descriptor: GpuDescriptorHandle
)

set a decriptor table in the graphics root signature

fn set_compute_rootsig(&mut self, rootsig: &RootSig)

set the compute root signature. TODO: distince root signatures

fn set_compute_root_constant(
&mut self,
param_index: u32,
value: u32,
param_offset: u32
)

set a constant in the compute root signature

fn set_compute_root_constants(
&mut self,
param_index: u32,
values: &[u32],
param_offset: u32
)

set a group of constants in the compute root signature

fn set_compute_root_cbv(&mut self, param_index: u32, resource: &mut RawResource)

set a cbv in the compute root signature

fn set_compute_root_srv(&mut self, param_index: u32, resource: &mut RawResource)

set a srv in the compute root signature

fn set_compute_root_uav(&mut self, param_index: u32, resource: &mut RawResource)

set a uav in the compute root signature

fn set_compute_root_dt(
&mut self,
param_index: u32,
base_descriptor: GpuDescriptorHandle
)

set a decriptor table in the compute root signature

fn clear_dsv(
&mut self,
heap: &mut DsvHeap,
index: u32,
flags: DepthStencilClearFlags,
depth: f32,
stencil: u8,
rects: Option<&[Rect]>
)

record a command to clear the dsv. more info

fn clear_uav_f32(
&mut self,
heap: &mut CbvSrvUavHeap,
index: u32,
resource: &mut RawResource,
values: &[f32; 4],
rects: Option<&[Rect]>
)

record a command to clear uav. more info

fn clear_uav_u32(
&mut self,
heap: &mut CbvSrvUavHeap,
index: u32,
resource: &mut RawResource,
values: &[u32; 4],
rects: Option<&[Rect]>
)

record a command to clear uav. more info

fn clear_rtv(
&mut self,
heap: &mut RtvHeap,
index: u32,
values: &[f32; 4],
rects: Option<&[Rect]>
)

record clearing a rtv

fn copy_resource(&mut self, dst: &mut RawResource, src: &mut RawResource)

record a resource copy operation.

restrictions

must be different resources

must be the same type

must have identical dimensions

must have compatible formats

more

fn copy_buffer_region(
&mut self,
dst: &mut RawResource,
dst_offset: u64,
src: &mut RawResource,
src_offset: u64,
size: u64
)

record a buffer copy operation

fn copy_texture_region(
&mut self,
dst: TextureCopyLocation,
dstx: u32,
dsty: u32,
dstz: u32,
src: TextureCopyLocation,
src_box: &Box3u
)

record a texture copy operation. more info

fn dispatch(&mut self, x: u32, y: u32, z: u32)

execute a command list from a thread group. more info

fn execute_bundle(&mut self, bundle: &mut Self)

execute a bundle. more

fn ia_set_primitive_topology(&mut self, topology: PrimitiveTopology)

set primitive topology

fn ia_set_ibv(&mut self, ibv: &IndexBufferView)

set the index buffer

fn ia_set_vbvs(&mut self, start_slot: u32, vbvs: &[VertexBufferView])

set the vertex buffer

fn so_set_targets(&mut self, start_slot: u32, sovs: &[StreamOutputBufferView])

set the stream output buffer views

fn rs_set_scissors(&mut self, scissors: &[Rect])

set scissor rectangles

fn rs_set_viewports(&mut self, viewports: &[Viewport])

set viewports

fn om_set_blend_factor(&mut self, factors: [f32; 4])

set the blend factor

fn om_set_stencil_ref(&mut self, stencil_ref: u32)

set the stencil reference

fn om_set_rtv_dsv_continuous(
&mut self,
rtv_heap: &mut RtvHeap,
rtv_offset: u32,
num_rtvs: u32,
dsv_heap: &mut DsvHeap,
dsv_offset: u32
)

set rtv and dsvs

fn om_set_rtv_dsv_discontinuous(
&mut self,
rtvs: &[CpuDescriptorHandle],
dsv_heap: &mut DsvHeap,
dsv_offset: u32
)

set rtv and dsvs

fn resolve_ms(
&mut self,
dst: &mut RawResource,
dst_sub: u32,
src: &mut RawResource,
src_sub: u32,
format: DxgiFormat
)

resolve a multisampled resource into a non-MS resource. more

fn resource_barrier(&mut self, barriers: &ResourceBarrierBuilder)

synchronizaing multiple access to resources. more

fn reset(
&mut self,
alloc: &mut CommandAllocator,
initial_state: Option<&PipelineState>
) -> Result<(), WinError>

reset a command list back to the initial state. more

fn clear_state(&mut self, initial_state: Option<&PipelineState>)

reset the state of a direct command list back to the state it was in when created. initial_state has to match this state

fn set_pipeline_state(&mut self, state: Option<&PipelineState>)

set the pipeline state

fn close(&mut self) -> Result<(), WinError>

close the current recording