Trait redirect::command::CommandListCommon

pub trait CommandListCommon {
    fn set_graphics_rootsig(&mut self, rootsig: &RootSig);
    fn set_graphics_root_constant(
        &mut self, 
        param_index: u32, 
        value: u32, 
        param_offset: u32
    );
    fn set_graphics_root_constants(
        &mut self, 
        param_index: u32, 
        values: &[u32], 
        param_offset: u32
    );
    fn set_graphics_root_cbv(
        &mut self, 
        param_index: u32, 
        resource: &mut RawResource
    );
    fn set_graphics_root_srv(
        &mut self, 
        param_index: u32, 
        resource: &mut RawResource
    );
    fn set_graphics_root_uav(
        &mut self, 
        param_index: u32, 
        resource: &mut RawResource
    );
    fn set_graphics_root_dt<H: Into<D3D12_GPU_DESCRIPTOR_HANDLE>>(
        &mut self, 
        param_index: u32, 
        base_descriptor: H
    );
    fn set_compute_rootsig(&mut self, rootsig: &RootSig);
    fn set_compute_root_constant(
        &mut self, 
        param_index: u32, 
        value: u32, 
        param_offset: u32
    );
    fn set_compute_root_constants(
        &mut self, 
        param_index: u32, 
        values: &[u32], 
        param_offset: u32
    );
    fn set_compute_root_cbv(
        &mut self, 
        param_index: u32, 
        resource: &mut RawResource
    );
    fn set_compute_root_srv(
        &mut self, 
        param_index: u32, 
        resource: &mut RawResource
    );
    fn set_compute_root_uav(
        &mut self, 
        param_index: u32, 
        resource: &mut RawResource
    );
    fn set_compute_root_dt<H: Into<D3D12_GPU_DESCRIPTOR_HANDLE>>(
        &mut self, 
        param_index: u32, 
        base_descriptor: H
    );
    fn dispatch(&mut self, x: u32, y: u32, z: u32);
    fn ia_set_primitive_topology(&mut self, topology: PrimitiveTopology);
    fn ia_set_ibv(&mut self, ibv: &IndexBufferView);
    fn ia_set_vbvs(&mut self, start_slot: u32, vbvs: &[VertexBufferView]);
    fn om_set_blend_factor(&mut self, factors: [f32; 4]);
    fn om_set_stencil_ref(&mut self, stencil_ref: u32);
    fn om_set_rtv_dsv_continuous(
        &mut self, 
        rtv_heap: &mut RtvHeap, 
        rtv_offset: u32, 
        num_rtvs: u32, 
        dsv: CpuDsvHandle
    );
    fn om_set_rtv_dsv_discontinuous(
        &mut self, 
        rtvs: &[CpuRtvHandle], 
        dsv: CpuDsvHandle
    );
    fn set_pipeline_state(&mut self, state: Option<&PipelineState>);
}

Required Methods

fn set_graphics_rootsig(&mut self, rootsig: &RootSig)

set the graphics root signature.

fn set_graphics_root_constant(
&mut self,
param_index: u32,
value: u32,
param_offset: u32
)

set a constant in the graphics root signature

fn set_graphics_root_constants(
&mut self,
param_index: u32,
values: &[u32],
param_offset: u32
)

set a group of constants in the graphics root signature

fn set_graphics_root_cbv(
&mut self,
param_index: u32,
resource: &mut RawResource
)

set a cbv in the graphics root signature

fn set_graphics_root_srv(
&mut self,
param_index: u32,
resource: &mut RawResource
)

set a srv in the graphics root signature

fn set_graphics_root_uav(
&mut self,
param_index: u32,
resource: &mut RawResource
)

set a uav in the graphics root signature

fn set_graphics_root_dt<H: Into<D3D12_GPU_DESCRIPTOR_HANDLE>>(
&mut self,
param_index: u32,
base_descriptor: H
)

set a decriptor table in the graphics root signature TODO: double check

fn set_compute_rootsig(&mut self, rootsig: &RootSig)

set the compute root signature. TODO: distince root signatures

fn set_compute_root_constant(
&mut self,
param_index: u32,
value: u32,
param_offset: u32
)

set a constant in the compute root signature

fn set_compute_root_constants(
&mut self,
param_index: u32,
values: &[u32],
param_offset: u32
)

set a group of constants in the compute root signature

fn set_compute_root_cbv(&mut self, param_index: u32, resource: &mut RawResource)

set a cbv in the compute root signature

fn set_compute_root_srv(&mut self, param_index: u32, resource: &mut RawResource)

set a srv in the compute root signature

fn set_compute_root_uav(&mut self, param_index: u32, resource: &mut RawResource)

set a uav in the compute root signature

fn set_compute_root_dt<H: Into<D3D12_GPU_DESCRIPTOR_HANDLE>>(
&mut self,
param_index: u32,
base_descriptor: H
)

set a decriptor table in the compute root signature TODO: double check

fn dispatch(&mut self, x: u32, y: u32, z: u32)

execute a command list from a thread group. more info

fn ia_set_primitive_topology(&mut self, topology: PrimitiveTopology)

set primitive topology

fn ia_set_ibv(&mut self, ibv: &IndexBufferView)

set the index buffer

fn ia_set_vbvs(&mut self, start_slot: u32, vbvs: &[VertexBufferView])

set the vertex buffer

fn om_set_blend_factor(&mut self, factors: [f32; 4])

set the blend factor

fn om_set_stencil_ref(&mut self, stencil_ref: u32)

set the stencil reference

fn om_set_rtv_dsv_continuous(
&mut self,
rtv_heap: &mut RtvHeap,
rtv_offset: u32,
num_rtvs: u32,
dsv: CpuDsvHandle
)

set rtv and dsvs

fn om_set_rtv_dsv_discontinuous(
&mut self,
rtvs: &[CpuRtvHandle],
dsv: CpuDsvHandle
)

set rtv and dsvs

fn set_pipeline_state(&mut self, state: Option<&PipelineState>)

set the pipeline state