Trait redirect::command::CommandListWithHeap

pub trait CommandListWithHeap {
    fn set_graphics_root_dt<H: Into<D3D12_GPU_DESCRIPTOR_HANDLE>>(
        &mut self, 
        param_index: u32, 
        base_descriptor: H
    );
    fn set_compute_root_dt<H: Into<D3D12_GPU_DESCRIPTOR_HANDLE>>(
        &mut self, 
        param_index: u32, 
        base_descriptor: H
    );
}

common methods for a command list with bounded descriptor heaps

Required Methods

fn set_graphics_root_dt<H: Into<D3D12_GPU_DESCRIPTOR_HANDLE>>(
&mut self,
param_index: u32,
base_descriptor: H
)

set a decriptor table in the graphics root signature TODO: double check

fn set_compute_root_dt<H: Into<D3D12_GPU_DESCRIPTOR_HANDLE>>(
&mut self,
param_index: u32,
base_descriptor: H
)

set a decriptor table in the compute root signature TODO: double check