Module redirect::descriptor::desc

descriptor description boilerplates

Structs

CbvDesc
describes constant buffer view

DsvDesc
describes a depth stencil view

DsvFlags
dsv misc flags. specifying read only would allow more than one dsv to be bound to the pipeline (on a same resource) simultaneously

DsvTex1DArrayDesc

DsvTex1DDesc

DsvTex2DArrayDesc

DsvTex2DDesc

DsvTex2DMsArrayDesc

RtvBufferDesc

RtvDesc
describes a render target view

RtvTex1DArrayDesc

RtvTex1DDesc

RtvTex2DArrayDesc

RtvTex2DDesc

RtvTex2DMsArrayDesc

RtvTex3DDesc

SamplerDesc
describes a sampler

Shader4ComponentMapping

ShaderComponentMapping
specifies how memory gets routed by a srv

SrvBufferDesc

SrvDesc
describes a shader resource view

SrvDescBindHelper
helper struct for ffi, not intended for application user TODO: remove from public interface

SrvTex1DArrayDesc

SrvTex1DDesc

SrvTex2DArrayDesc

SrvTex2DDesc

SrvTex2DMsArrayDesc

SrvTex3DDesc

SrvTexCubeArrayDesc

SrvTexCubeDesc

UavBufferDesc

UavDesc
describes an unordered access view

UavDescBindHelper
helper struct for ffi, not intended for application user TODO: remove from public interface

UavTex1DArrayDesc

UavTex1DDesc

UavTex2DArrayDesc

UavTex2DDesc

UavTex3DDesc

Enums

DsvDimension

RtvDimension

SrvDimension

UavDimension

Constants

DSV_FLAG_NONE

DSV_FLAG_READ_ONLY_DEPTH

DSV_FLAG_READ_ONLY_STENCIL

SHADER_COMPONENT_MAPPING_FORCE_VALUE_0
indicates forcing the resulting value to 0

SHADER_COMPONENT_MAPPING_FORCE_VALUE_1
indicates forcing the resulting value to 0x1 or 1.0f

SHADER_COMPONENT_MAPPING_FROM_MEMORY_COMPONENT_0
indicates return component 0, i.e. R in RGBA

SHADER_COMPONENT_MAPPING_FROM_MEMORY_COMPONENT_1
indicates return component 1, i.e. G in RGBA

SHADER_COMPONENT_MAPPING_FROM_MEMORY_COMPONENT_2
indicates return component 2, i.e. B in RGBA

SHADER_COMPONENT_MAPPING_FROM_MEMORY_COMPONENT_3
indicates return component 3, i.e. A in RGBA

Type Definitions

ComparisonFunc