Module redirect::descriptor::heap

a descriptor heap is where the descriptors resides on

Structs

CpuCsuHandle

CpuDsvHandle

CpuRtvHandle

CpuSamplerHandle

CsuHeapNsv
a heap that can hold cbv, srv and uavs, that is not shader visible.

CsuHeapSv
a heap that can hold cbv, srv and uavs, that is shader visible.

DescriptorHeapBuilder
descriptor heap builder struct

DsvHeap
a heap that can hold depth stencil views

GpuCsuHandle

GpuDsvHandle

GpuRtvHandle

GpuSamplerHandle

RtvHeap
a heap that can hold render target views

SamplerHeapNsv
a heap that can hold samplers, that is not shader visible

SamplerHeapSv
a heap that can hold samplers, that is shader visible

Traits

CsuHeap

DescriptorHeap
represents a descriptor heap

DescriptorHeapCopy
represents a descriptor heap that can be used as a copy source

SamplerHeap