Struct redirect::descriptor::heap::DsvHeap

pub struct DsvHeap {
    pub ptr: ComPtr<ID3D12DescriptorHeap>,
    // some fields omitted
}

a heap that can hold depth stencil views

Fields
ptr: ComPtr<ID3D12DescriptorHeap>
Methods

impl DsvHeap

fn create_dsv(
&mut self,
device: &mut Device,
resource: Option<&mut RawResource>,
desc: Option<&DsvDesc>,
index: u32
)

create a depth stencil view on the given resource. A None desc means to create a default view if possible