Trait redirect::descriptor::heap::DescriptorHeap

pub trait DescriptorHeap {
    type CpuHandle;
    type GpuHandle;
    fn get_type(&self) -> D3D12_DESCRIPTOR_HEAP_TYPE;
    fn get_cpu_handle(&mut self, offset: u32) -> Self::CpuHandle;
    fn get_gpu_handle(&mut self, offset: u32) -> Self::GpuHandle;
    fn copy_descriptors_to(
        &mut self, 
        dst: &mut Self, 
        device: &mut Device, 
        src_offset: u32, 
        dst_offset: u32, 
        num_descriptors: u32
    );
    fn len(&self) -> u32;
    fn get_handle_increment_size(&self) -> u32;
}

represents a descriptor heap

Associated Types

type CpuHandle

type GpuHandle

Required Methods

fn get_type(&self) -> D3D12_DESCRIPTOR_HEAP_TYPE

get type of the heap

fn get_cpu_handle(&mut self, offset: u32) -> Self::CpuHandle

get Cpu handle of a descriptor at offset on the heap

fn get_gpu_handle(&mut self, offset: u32) -> Self::GpuHandle

get Gpu handle of a descriptor at offset on the heap

fn copy_descriptors_to(
&mut self,
dst: &mut Self,
device: &mut Device,
src_offset: u32,
dst_offset: u32,
num_descriptors: u32
)

perform immediate copy of a slice of descriptors on CPU side through the given device

fn len(&self) -> u32

get the number of descriptors this heap can hold

fn get_handle_increment_size(&self) -> u32

get the handle incremental size of this heap