Trait redirect::descriptor::heap::DescriptorHeapCopy

pub trait DescriptorHeapCopy<T: DescriptorHeap>: DescriptorHeap {
    fn copy_descriptors_to(
        &mut self, 
        dst: &mut T, 
        device: &mut Device, 
        src_offset: u32, 
        dst_offset: u32, 
        num_descriptors: u32
    ) { ... }
}

represents a descriptor heap that can be used as a copy source

Provided Methods

fn copy_descriptors_to(
&mut self,
dst: &mut T,
device: &mut Device,
src_offset: u32,
dst_offset: u32,
num_descriptors: u32
)

perform immediate copy of a slice of descriptors on CPU timeslice through the given device. self must not be shader visible.