Struct redirect::device::Device

pub struct Device {
    pub ptr: ComPtr<ID3D12Device>,
}

a 3D display adapter

Fields
ptr: ComPtr<ID3D12Device>
Methods

impl Device

fn new(
adapter: Option<&Adapter>,
level: FeatureLevel
) -> Result<Device, WinError>

attempt to create a device from the given adapter and feature level. None means the default adapter would be used.

fn create_graphics_pso<'a>(
&mut self,
pso: &mut GraphicsPipelineStateBuilder<'a>
) -> Result<PipelineState, WinError>

attempts to create a graphics pipeline state object

fn create_root_sig(
&mut self,
node_mask: u32,
desc_blob: &RootSigDescBlob
) -> Result<RootSig, WinError>

attempts to create a root signature from a description blob

fn create_command_queue(
&mut self,
desc: &CommandQueueDesc
) -> Result<CommandQueue, WinError>

attempts to create a command queue with given description

fn create_direct_command_allocator(
&mut self
) -> Result<DirectCommandAllocator, WinError>

attempts to create a direct command allocator

fn create_bundle_allocator(
&mut self
) -> Result<BundleCommandAllocator, WinError>

attempts to create a bundle command allocator

fn create_heap(&mut self, desc: &HeapDesc) -> Result<Heap, WinError>

attempts to create a heap

fn create_committed_resource(
&mut self,
heap_properties: &HeapProperties,
heap_flags: HeapFlags,
desc: &ResourceDesc,
initial_state: ResourceStates
) -> Result<CommittedResource, WinError>

attempts to create a committed resource

fn create_fence(
&mut self,
initial_value: u64,
flags: FenceFlags
) -> Result<Fence, WinError>

attempts to create a fence

fn get_resource_alloc_info(
&mut self,
desc: &ResourceDesc,
visible_mask: u32
) -> ResourceAllocInfo

get resource allocation info from the resource description

fn create_placed_resource(
&mut self,
heap: &mut Heap,
heap_offset: u64,
desc: &ResourceDesc,
initial_state: ResourceStates
) -> Result<PlacedResource, WinError>

attempts to create a placed resource

fn create_cbv_srv_uav_heap(
&mut self,
builder: &DescriptorHeapBuilder
) -> Result<CbvSrvUavHeap, WinError>

fn create_rtv_heap(
&mut self,
builder: &DescriptorHeapBuilder
) -> Result<RtvHeap, WinError>

fn create_dsv_heap(
&mut self,
builder: &DescriptorHeapBuilder
) -> Result<DsvHeap, WinError>

fn create_sampler_heap(
&mut self,
builder: &DescriptorHeapBuilder
) -> Result<SamplerHeap, WinError>

fn create_direct_command_list<'a>(
&mut self,
node_mask: u32,
alloc: &'a mut DirectCommandAllocator,
initial_state: Option<&'a PipelineState>
) -> Result<DirectCommandListRecording<'a>, WinError>

fn create_bundle<'a>(
&mut self,
node_mask: u32,
alloc: &'a mut BundleCommandAllocator,
initial_state: Option<&PipelineState>
) -> Result<BundleRecording<'a>, WinError>