Struct redirect::event::Event

pub struct Event { /* fields omitted */ }

RAII Win32 event

Methods

impl Event

pub unsafe fn from_handle(handle: HANDLE) -> Event

Obtain the handle and gain exclusive ownership to it.

Safety

Caller must ensure that handle is a valid event, and that it would not be used elsewhere afterwards.

pub fn new<'a>(flags: Flag, access: AccessRight) -> Result<Event, u32>

pub fn reset(&self) -> Result<(), u32>

Reset a manually reset event

pub fn set(&self) -> Result<(), u32>

set the event

pub fn wait(&self) -> Result<(), u32>

wait for this event to be set

pub fn wait_for(&self, milli_sec: u32) -> Result<(), u32>

wait for milli_sec ms

pub fn get(&self) -> HANDLE

get the event raw handle