Module redirect::factory

this module contains Factory, the API entry point, along with several other fundamental structures such as Adapter and Output.

Structs

Adapter
a display subsystem

AdapterDesc
adapter description

AdapterFlags
adapter flags

AdapterIter
iterator returned by a factory to retrieve available adapters

Factory
dxgi API entry point

Output
a display output, such as a monitor

OutputDesc

OutputIter
iterator returned by an adapter to retrieve available outputs

RotationMode
rotation mode for the monitor

VideoMemInfo
video memory informaiton

Constants

ADAPTER_FLAG_NONE

ADAPTER_FLAG_REMOTE
reserved flag

ADAPTER_FLAG_SOFTWARE
software adapter

ROTATION_MODE_IDENTITY

ROTATION_MODE_ROTATE90

ROTATION_MODE_ROTATE180

ROTATION_MODE_ROTATE270

ROTATION_MODE_UNSPECIFIED

Traits

HwndProvider
a provider for HWND