Module redirect::pipeline

describes graphics and compute pipelines

Modules

blend
blend state

ds
depth stencil states

ia
input element layout

rasterizer
defines the rasterizer state

rootsig
describes what kind of resources are to be bound to the pipeline.

sampler
texture sampling schemes

so
stream output descriptions

Structs

ComparisonFunc
comparison options, specifying under which circumstance the comparison passes

ComputePipelineState
a pipeline state object

ComputePipelineStateBuilder

ComputePipelineStateCache
a pipeline state cached blob

GraphicsPipelineState
a graphics pipeline state object

GraphicsPipelineStateBuilder
graphics pso builder

GraphicsPipelineStateCache
a graphics pipeline state cached blob

PipelineStateFlags

Constants

COMPARISON_FUNC_ALWAYS

COMPARISON_FUNC_EQUAL

COMPARISON_FUNC_GREATER

COMPARISON_FUNC_GREATER_EQUAL

COMPARISON_FUNC_LESS

COMPARISON_FUNC_LESS_EQUAL

COMPARISON_FUNC_NEVER

COMPARISON_FUNC_NOT_EQUAL

PIPELINE_STATE_FLAG_NONE

PIPELINE_STATE_FLAG_TOOL_DEBUG

Traits

PipelineState
a pipeline state object

Type Definitions

SampleDesc