Module redirect::pipeline::rootsig

describes what kind of resources are to be bound to the pipeline.

Structs

DescriptorRange
descriptor range

DescriptorRangeType
type of a descriptor range

RootParam
describes a root parameter

RootSig
a root signature

RootSigBuilder
builder for a root signature

RootSigDescBlob
a serialized root signature description blob

RootSigFlags
misc flags for a root signature

ShaderVisibility
specifies which shader can access content of a given root parameter

Enums

RootParamType
specifies a type of root parameter

Constants

DESCRIPTOR_RANGE_TYPE_CBV

DESCRIPTOR_RANGE_TYPE_SAMPLER

DESCRIPTOR_RANGE_TYPE_SRV

DESCRIPTOR_RANGE_TYPE_UAV

ROOT_SIGNATURE_FLAG_ALLOW_INPUT_ASSEMBLER_INPUT_LAYOUT

ROOT_SIGNATURE_FLAG_ALLOW_STREAM_OUTPUT

ROOT_SIGNATURE_FLAG_DENY_DOMAIN_SHADER_ROOT_ACCESS

ROOT_SIGNATURE_FLAG_DENY_GEOMETRY_SHADER_ROOT_ACCESS

ROOT_SIGNATURE_FLAG_DENY_HULL_SHADER_ROOT_ACCESS

ROOT_SIGNATURE_FLAG_DENY_PIXEL_SHADER_ROOT_ACCESS

ROOT_SIGNATURE_FLAG_DENY_VERTEX_SHADER_ROOT_ACCESS

ROOT_SIGNATURE_FLAG_NONE

SHADER_VISIBILITY_ALL

SHADER_VISIBILITY_DOMAIN

SHADER_VISIBILITY_GEOMETRY

SHADER_VISIBILITY_HULL

SHADER_VISIBILITY_PIXEL

SHADER_VISIBILITY_VERTEX