Enum redirect::pipeline::rootsig::RootParamType

pub enum RootParamType {
    DescriptorTable {
        descriptor_ranges: SmallVec<[DescriptorRange; 4]>,
    },
    Cbv {
        shader_register: u32,
        register_space: u32,
    },
    Srv {
        shader_register: u32,
        register_space: u32,
    },
    Uav {
        shader_register: u32,
        register_space: u32,
    },
    Constant {
        shader_register: u32,
        register_space: u32,
        num_32bit_values: u32,
    },
}

specifies a type of root parameter

Variants
DescriptorTable

a collection of descriptor ranges, appearing in sequence in a descriptor heap

Fields of DescriptorTable

descriptor_ranges: SmallVec<[DescriptorRange; 4]>

an array of descriptor ranges

Cbv

cbv descriptor inlined in the signature

Fields of Cbv

shader_register: u32

the shader register

register_space: u32

the register space

Srv

srv descriptor inlined in the signature

Fields of Srv

shader_register: u32

the shader register

register_space: u32

the register space

Uav

uav descriptor inlined in the signature

Fields of Uav

shader_register: u32

the shader register

register_space: u32

the register space

Constant

constants inlined in the signature that appear in shaders as one constant buffer

Fields of Constant

shader_register: u32

shader register of this constant

register_space: u32

register space of this constant

num_32bit_values: u32

number of 32bit values in this constant slot