Module redirect::pipeline::sampler

texture sampling schemes

Structs

BorderColor
border colors

Filter
filtering options for texture sampling. more info

StaticSamplerDesc
describes a static sampler

TextureAddressMode
texture addressing modes when coordinates outside normalized boundary

Constants

BORDER_COLOR_OPAQUE_BLACK

BORDER_COLOR_OPAQUE_WHITE

BORDER_COLOR_TRANSPARENT_BLACK

FILTER_ANISOTROPIC

FILTER_COMPARISON_ANISOTROPIC

FILTER_COMPARISON_MIN_LINEAR_MAG_MIP_POINT

FILTER_COMPARISON_MIN_LINEAR_MAG_POINT_MIP_LINEAR

FILTER_COMPARISON_MIN_MAG_LINEAR_MIP_POINT

FILTER_COMPARISON_MIN_MAG_MIP_LINEAR

FILTER_COMPARISON_MIN_MAG_MIP_POINT

FILTER_COMPARISON_MIN_MAG_POINT_MIP_LINEAR

FILTER_COMPARISON_MIN_POINT_MAG_LINEAR_MIP_POINT

FILTER_COMPARISON_MIN_POINT_MAG_MIP_LINEAR

FILTER_MAXIMUM_ANISOTROPIC

FILTER_MAXIMUM_MIN_LINEAR_MAG_MIP_POINT

FILTER_MAXIMUM_MIN_LINEAR_MAG_POINT_MIP_LINEAR

FILTER_MAXIMUM_MIN_MAG_LINEAR_MIP_POINT

FILTER_MAXIMUM_MIN_MAG_MIP_LINEAR

FILTER_MAXIMUM_MIN_MAG_MIP_POINT

FILTER_MAXIMUM_MIN_MAG_POINT_MIP_LINEAR

FILTER_MAXIMUM_MIN_POINT_MAG_LINEAR_MIP_POINT

FILTER_MAXIMUM_MIN_POINT_MAG_MIP_LINEAR

FILTER_MINIMUM_ANISOTROPIC

FILTER_MINIMUM_MIN_LINEAR_MAG_MIP_POINT

FILTER_MINIMUM_MIN_LINEAR_MAG_POINT_MIP_LINEAR

FILTER_MINIMUM_MIN_MAG_LINEAR_MIP_POINT

FILTER_MINIMUM_MIN_MAG_MIP_LINEAR

FILTER_MINIMUM_MIN_MAG_MIP_POINT

FILTER_MINIMUM_MIN_MAG_POINT_MIP_LINEAR

FILTER_MINIMUM_MIN_POINT_MAG_LINEAR_MIP_POINT

FILTER_MINIMUM_MIN_POINT_MAG_MIP_LINEAR

FILTER_MIN_LINEAR_MAG_MIP_POINT

FILTER_MIN_LINEAR_MAG_POINT_MIP_LINEAR

FILTER_MIN_MAG_LINEAR_MIP_POINT

FILTER_MIN_MAG_MIP_LINEAR

FILTER_MIN_MAG_MIP_POINT

FILTER_MIN_MAG_POINT_MIP_LINEAR

FILTER_MIN_POINT_MAG_LINEAR_MIP_POINT

FILTER_MIN_POINT_MAG_MIP_LINEAR

TEXTURE_ADDRESS_MODE_BORDER
set to a "border color"

TEXTURE_ADDRESS_MODE_CLAMP
clamp to values at normalized boundary

TEXTURE_ADDRESS_MODE_MIRROR
filp at every integer junction

TEXTURE_ADDRESS_MODE_MIRROR_ONCE
take the absolution value of texture coordinates, then clamp to the boundary

TEXTURE_ADDRESS_MODE_WRAP
tile at every integer junction, essentially repeating the textures