Module redirect::resource::barrier

resource barriers

Structs

ResourceAliasingBarrier
describes the transition between usage of two different resources mapping into the same heap location.

ResourceBarrier
resource barriers

ResourceBarrierFlags
misc resource barrier flags. more

ResourceBarriersBuilder
resource barrier builder

ResourceTransitionBarrier
represents a transition of subresource between different usages

ResourceUavBarrier
describes a barrier in which all uav access to a resource must complete before subsequent uav accesses can begin

Enums

ResourceBarrierType

Constants

RESOURCE_BARRIER_FLAG_BEGIN_ONLY
Marks the start of a transition, putting the resource in a temporary no-access condition.

RESOURCE_BARRIER_FLAG_END_ONLY
Marks the end of a transition, restore the resource to the state after transition.

RESOURCE_BARRIER_FLAG_NONE
The default, an atomic barrier.