Module redirect::resource::description

attributes used to describe a resource

Structs

ResourceAlignment
alignment of the resource

ResourceDesc
resource description

ResourceDimension
dimension i.e. type of the resource

ResourceFlags
misc flags for resources. more info

TextureLayout
texture layout

Constants

RESOURCE_4KB_ALIGNED
4kb aligned

RESOURCE_4MB_ALIGNED
4mb aligned

RESOURCE_64KB_ALIGNED
64kb aligned

RESOURCE_DEFAULT_ALIGNED
4mb for msaa textures, 64kb for everything else. This is the deefault.

RESOURCE_DIMENSION_BUFFER

RESOURCE_DIMENSION_TEXTURE1D

RESOURCE_DIMENSION_TEXTURE2D

RESOURCE_DIMENSION_TEXTURE3D

RESOURCE_DIMENSION_UNKNOWN

RESOURCE_FLAG_ALLOW_CROSS_ADAPTER
Allows the resource to be used for cross-adapter data, as well as the same features enabled by AllowSimultaneousAccess. Must be used with heaps with SharedCrossAdapter and not AllowDisplay.

RESOURCE_FLAG_ALLOW_DEPTH_STENCIL
Allow dsv on the resource, allow resource state transition to DepthWrite/DepthRead. - Texture format must support depth stencil capability at the current feature level. - Cannot be used with: - Buffer, - 4kb resource alignment, - AllowRenderTarget, - AllowUnorderedAccess, - AllowSimultaneousAccess, - 64KbStandardSwizzle, - RowMajor, - DenyRtDsTextures, - AllowDisplay

RESOURCE_FLAG_ALLOW_RENDER_TARGET
Allow the resource to be used as render target. - Should be used with format supporting render target capabilities At current feature level. - Can't be used in conjunction with RowMajorLayout when ... (see doc for more) - Can't be useed with 4kb resource alignment, or AllowDepthStencil, pr DenyRtDsTextures

RESOURCE_FLAG_ALLOW_SIMULTANEOUS_ACCESS
Allows a resource to be simultaneously accessed by multiple different queues, devices or processes - Cannot be used with Buffer, coz bbuffer always have this property..>.< - Cannot be used with MSAA

RESOURCE_FLAG_ALLOW_UNORDERED_ACCESS
Allow uav on the resource, allow resource state transition to UnorderedAccess. - Texture format must support unordered access capability at the current feature level. - Cannot be used with MSAA and ... (see doc for more)

RESOURCE_FLAG_DENY_SHADER_RESOURCE
Disallows a srv to be created for the resource, as well as disables the resource to transition into the state of SharedResources. Must be used with AllowDepthStencil

RESOURCE_FLAG_NONE
Default flag

TEXTURE_LAYOUT_64KB_STANDARD_SWIZZLE

TEXTURE_LAYOUT_64KB_UNDEFINED_SWIZZLE

TEXTURE_LAYOUT_ROW_MAJOR
data for the texture is stored in row-major order. only the following texture properties are supported:

TEXTURE_LAYOUT_UNKNOWN
adapter-dependent layout. driver choose optimal layout during resource creation