Module redirect::resource::heap::description

Types to describe a heap

Structs

HeapAlignment
heap alignment

HeapDesc
description of a heap

HeapFlags
misc heap options. more info

HeapProperties
describes heap properties

HeapType
heap type.

MemoryPoolPreference
memory pool preference. more info

PageProperty
cpu page properties.

Constants

HEAP_ALIGNMENT_DEFAULT
alias for 64kb, the default.

HEAP_ALIGNMENT_DEFAULT_MSAA_RESOURCE_PLACEMENT
4mb aligned. MSAA resource heap must use this alignment.

HEAP_ALIGNMENT_DEFAULT_RESOURCE_PLACEMENT
64kb aligned.

HEAP_FLAG_ALLOW_ALL_BUFFERS_AND_TEXTURES

HEAP_FLAG_ALLOW_DISPLAY
the heap can contain swapchain surfaces

HEAP_FLAG_ALLOW_ONLY_BUFFERS

HEAP_FLAG_ALLOW_ONLY_NON_RT_DS_TEXTURES

HEAP_FLAG_ALLOW_ONLY_RT_DS_TEXTURES

HEAP_FLAG_ALLOW_WRITE_WATCH
allow tools to support MEM_WRITE_WATCH

HEAP_FLAG_DENY_BUFFERS
the heap isn't allowed to contain buffers

HEAP_FLAG_DENY_NON_RT_DS_TEXTURES
the heap can't contain textures without ALLOW_RENDER_TARGET or ALLOW_DEPTH_STENCIL flags

HEAP_FLAG_DENY_RT_DS_TEXTURES
the heap can't store render target or depth stencil textures

HEAP_FLAG_HARDWARE_PROTECTED
unsupported

HEAP_FLAG_NONE
The default, no options specified.

HEAP_FLAG_SHARED
a shared heap

HEAP_FLAG_SHARED_CROSS_ADAPTER
the heap can be shored across adapters

HEAP_TYPE_CUSTOM
Custom heap for advanced usage.

HEAP_TYPE_DEFAULT
GPU RW, no CPU access. This is the default heap type.

HEAP_TYPE_READBACK
Optimal for CPU write. Best for GPU write-once, CPU readable data. Resources in this heap must be created with COPY_DEST state, and cannot be changed away from this.

HEAP_TYPE_UPLOAD
Optimal for CPU write. Best for CPU write-once, GPU read-once data. Resources in this heap must be created with GENERATE_READ state, and cannot be changed away.

MEMORY_POOL_L0

MEMORY_POOL_L1

MEMORY_POOL_UNKNOWN
The default pool preference.

PAGE_PROPERTY_NOT_AVAILABLE
The CPU cannot access the heap, thus no property available.

PAGE_PROPERTY_UNKNOWN
The default cpu page property.

PAGE_PROPERTY_WRITE_BACK

PAGE_PROPERTY_WRITE_COMBINE