Trait redirect::resource::traits::Buffer

pub unsafe trait Buffer: Resource {
    fn get_size(&self) -> u64;

    fn is_compatible_with(&self, buffer_slice: BufferSlice) -> bool { ... }

    fn create_cbv<DH: CsuHeap>(
        &mut self, 
        device: &mut Device, 
        csu_heap: &mut DH, 
        index: u32
    ) { ... }

    fn create_srv<DH: CsuHeap, B: AllowShaderResource + Buffer>(
        buf: &mut B, 
        device: &mut Device, 
        csu_heap: &mut DH, 
        index: u32, 
        slice: BufferSlice
    ) { ... }

    fn create_uav<DH: CsuHeap, B: AllowUnorderedAccess + Buffer>(
        buf: &mut B, 
        device: &mut Device, 
        csu_heap: &mut DH, 
        index: u32, 
        slice: BufferSlice
    ) { ... }

    fn create_uav_with_counter<DH: CsuHeap, CounterBuf: Buffer, B: AllowUnorderedAccess + Buffer>(
        buf: &mut B, 
        device: &mut Device, 
        csu_heap: &mut DH, 
        index: u32, 
        slice: BufferSlice, 
        counter: &mut CounterBuf, 
        counter_offset: u64
    ) { ... }

    fn create_vbv(&mut self, size: u32, stride: u32) -> VertexBufferView { ... }

    fn create_ibv_u32(&mut self, size: u32) -> IndexBufferView { ... }

    fn create_ibv_u16(&mut self, size: u32) -> IndexBufferView { ... }
}

a buffer

Required Methods

fn get_size(&self) -> u64

get the size of the buffer

Provided Methods

fn is_compatible_with(&self, buffer_slice: BufferSlice) -> bool

Check if this buffer can hold the given buffer_slice

fn create_cbv<DH: CsuHeap>(
&mut self,
device: &mut Device,
csu_heap: &mut DH,
index: u32
)

create a cbv for this buffer on csu_heap at index

fn create_srv<DH: CsuHeap, B: AllowShaderResource + Buffer>(
buf: &mut B,
device: &mut Device,
csu_heap: &mut DH,
index: u32,
slice: BufferSlice
)

Create a srv for this buffer on csu_heap at index with the given slice TODO: verify, raw buffer support

fn create_uav<DH: CsuHeap, B: AllowUnorderedAccess + Buffer>(
buf: &mut B,
device: &mut Device,
csu_heap: &mut DH,
index: u32,
slice: BufferSlice
)

Create a uav for this buffer on csu_heap at index with the given slice TODO: verify

fn create_uav_with_counter<DH: CsuHeap, CounterBuf: Buffer, B: AllowUnorderedAccess + Buffer>(
buf: &mut B,
device: &mut Device,
csu_heap: &mut DH,
index: u32,
slice: BufferSlice,
counter: &mut CounterBuf,
counter_offset: u64
)

Create a uav for this buffer on csu_heap at index with the given slice and counter TODO: verify

fn create_vbv(&mut self, size: u32, stride: u32) -> VertexBufferView

Create a vbv for this buffer

fn create_ibv_u32(&mut self, size: u32) -> IndexBufferView

Create an ibv for this buffer

fn create_ibv_u16(&mut self, size: u32) -> IndexBufferView

Create an ibv for this buffer