Trait redirect::resource::traits::Texture

pub unsafe trait Texture: Resource {
    fn get_format(&mut self) -> DxgiFormat { ... }

    fn create_srv<DH: CsuHeap, T: AllowShaderResource + Texture>(
        tex: &mut T, 
        device: &mut Device, 
        csu_heap: &mut DH, 
        index: u32
    ) { ... }

    fn create_uav<DH: CsuHeap, T: AllowUnorderedAccess + Texture>(
        tex: &mut T, 
        device: &mut Device, 
        csu_heap: &mut DH, 
        index: u32
    ) { ... }

    fn create_rtv<T: AllowRenderTarget + Texture>(
        tex: &mut T, 
        device: &mut Device, 
        heap: &mut RtvHeap, 
        index: u32
    ) { ... }

    fn create_dsv<T: AllowDepthStencil + Texture>(
        tex: &mut T, 
        device: &mut Device, 
        heap: &mut DsvHeap, 
        index: u32
    ) { ... }
}

a texture

Provided Methods

fn get_format(&mut self) -> DxgiFormat

get the format of the texel

fn create_srv<DH: CsuHeap, T: AllowShaderResource + Texture>(
tex: &mut T,
device: &mut Device,
csu_heap: &mut DH,
index: u32
)

Create a srv for this texture on csu_heap at index TODO: verify, SrvTex1DDesc support

fn create_uav<DH: CsuHeap, T: AllowUnorderedAccess + Texture>(
tex: &mut T,
device: &mut Device,
csu_heap: &mut DH,
index: u32
)

Create a srv for this texture on csu_heap at index TODO: verify, UavTex1DDesc support

fn create_rtv<T: AllowRenderTarget + Texture>(
tex: &mut T,
device: &mut Device,
heap: &mut RtvHeap,
index: u32
)

Create a rtv for this texture on heap at index TODO: RtvTex1DDesc support

fn create_dsv<T: AllowDepthStencil + Texture>(
tex: &mut T,
device: &mut Device,
heap: &mut DsvHeap,
index: u32
)

Create a dsv for this texture on heap at index TODO: DsvTex1DDesc support