Module redirect::swapchain

link between the graphics API and the target surface

Structs

AlphaMode
transparency behavior of a surface

FullScreenDesc
optional description of a fullsceen swapchain

ModeScaling
scaling behavior for an image on a monitor

PresentFlags
options for frame presentation

SampleDesc
multi-sampling scheme description. Default to count 1 and quality 0, representing no aa

Scaling
scaling behavor when the back buffer got presented.

ScanlineOrder
method the raster uses to create an image on the surface

SwapChain
link between the graphics API and the target surface

SwapChainDesc
description of a swapchain

SwapChainFlags
misc flags for swapchain behavior

SwapChainResizeDesc
parameters for swapchain resizing

SwapEffect
presentation model, as well as how the back buffer would be handled after calling swapchain.present(). more info

Constants

ALPHA_MODE_IGNORE
alpha channel would be ignored

ALPHA_MODE_PREMULTIPLIED
each color channel is premultiplied by the alpha value

ALPHA_MODE_STRAIGHT
each color channel is not premultiplied by the alpha value

ALPHA_MODE_UNSPECIFIED
transparency behavior is not specified

MODE_SCALING_CENTERED

MODE_SCALING_STRETCHED

MODE_SCALING_UNSPECIFIED

PRESENT_FLAG_ALLOW_TEARING
allow tearing for variable refresh rate displays.

PRESENT_FLAG_DO_NOT_SEQUENCE
present a frame from current buffer to the output. this flag allows vsync instead of typical sequencing

PRESENT_FLAG_DO_NOT_WAIT
make the invocation fail if the calling thread would be blocked

PRESENT_FLAG_NONE
present a frame from each buffer (starting from the current one) to the output

PRESENT_FLAG_RESTART
make the runtime discard outstanding queued frames

PRESENT_FLAG_RESTRICT_TO_OUTPUT
indicates that presentation content will be shown only on the particular output. The content will not be visible on other outputs.

PRESENT_FLAG_STEREO_PREFER_RIGHT
stereo prefers right-eye viewing instead of right

PRESENT_FLAG_STEREO_TEMPORARY_MONO
Indicates that the presentation should use the left buffer as a mono buffer.

PRESENT_FLAG_TEST
don't present to the output. intended for use only when switching from idle

SCALING_ASPECT_RATIO_STRETCH
back buffer content would be scaled to fit the presentation target, while preserving the aspect ratio, centered with black borders

SCALING_NONE
back buffer content would appear without scaling, with top edge aligned with the presentation target.

SCALING_STRETCH
back buffer content would be scaled to fill the presentation target

SCANLINE_ORDER_LOWER_FIELD_FIRST
image is created beginning with the lower field

SCANLINE_ORDER_PROGRESSIVE
image is created from the first scanline to the last without skipping any

SCANLINE_ORDER_UNSPECIFIED

SCANLINE_ORDER_UPPER_FIELD_FIRST
image is created beginning with the upper field

SWAP_CHAIN_FLAG_ALLOW_MODE_SWITCH
allow switch between fullscreen and windowed with resize_target

SWAP_CHAIN_FLAG_ALLOW_TEARING
enable displays that support variable refresh rates to function properly when the application presents a swapchain tied to a full screen borderless window.

SWAP_CHAIN_FLAG_DISPLAY_ONLY
the presented content would only be avaiable for local display

SWAP_CHAIN_FLAG_FOREGROUND_LAYER
create a swapchain in the foreground layer for multi-plane rendering

SWAP_CHAIN_FLAG_FRAME_LATENCY_WAITABLE_OBJECT
ensure rendering does not begin while a frame is still being resented

SWAP_CHAIN_FLAG_FULLSCREEN_VIDEO

SWAP_CHAIN_FLAG_GDI_COMPATIBLE
allow get_dc on the 0th back buffer

SWAP_CHAIN_FLAG_HW_PROTECTED

SWAP_CHAIN_FLAG_NONE

SWAP_CHAIN_FLAG_NONPREROTATED
turn off fullscreen automatic rotation

SWAP_CHAIN_FLAG_RESTRICTED_CONTENT
OS would support creation only when driver and hardware protection is used?

SWAP_CHAIN_FLAG_RESTRICT_SHARED_RESOURCE_DRIVER

SWAP_CHAIN_FLAG_YUV_VIDEO

SWAP_EFFECT_DISCARD
bitblt, back buffer content would be discarded after presented

SWAP_EFFECT_FLIP_DISCARD
flip, back buffer content would be discared after presented, cannot be used with multisampling and partial presentation

SWAP_EFFECT_FLIP_SEQUENTIAL
flip, back buffer content would persist after presented, cannot be used with multisampling

SWAP_EFFECT_SEQUENTIAL
bitblt, back buffer content would persist after presented, cannot be used with multisampling