Struct redirect::swapchain::SwapChain

pub struct SwapChain { /* fields omitted */ }

link between the graphics API and the target surface

Methods

impl SwapChain

pub fn get_current_back_buffer_index(&mut self) -> u32

gets the index of this swapchain's current back buffer

pub fn get_buffer(&mut self, index: u32) -> Result<RawResource, WinError>

get buffer at the given index

pub fn resize_buffers(
&mut self,
params: SwapChainResizeDesc
) -> Result<(), WinError>

attemp to resize the back buffers with given parameters

pub fn get_source_size(&mut self) -> Result<(u32, u32), WinError>

get the source region size for the swap chain

pub fn set_source_size(
&mut self,
width: u32,
height: u32
) -> Result<(), WinError>

set the source region size for the swap chain

pub fn get_background_color(&mut self) -> Result<[f32; 4], WinError>

get the background color for the next present method of this swapchain

pub fn set_background_color(
&mut self,
r: f32,
g: f32,
b: f32,
a: f32
) -> Result<(), WinError>

change the background color for the next frame

pub fn get_desc(&mut self) -> Result<SwapChainDesc, WinError>

get description

pub fn get_fullscreen_desc(&mut self) -> Result<FullScreenDesc, WinError>

get fullscreen description

pub fn get_hwnd(&mut self) -> Result<HWND, WinError>

get the underlying HWMD handle for the swapchain object

pub fn get_last_present_count(&mut self) -> Result<u32, WinError>

get the number of times that Present or Present1 has been called.

pub fn present(
&mut self,
sync_interval: u32,
flags: PresentFlags
) -> Result<(), WinError>

present a rendered back buffer to the target output. sync_interval specifies how to synchronize presentation of a frame with the verticle blank. valid values include [0..4].