Struct redirect::swapchain::SwapChainDesc

#[repr(C)]

pub struct SwapChainDesc {
    pub width: u32,
    pub height: u32,
    pub format: DxgiFormat,
    pub stereo: Bool,
    pub sample_desc: SampleDesc,
    pub buffer_usage: Usage,
    pub buffer_count: u32,
    pub scaling: Scaling,
    pub swap_effect: SwapEffect,
    pub alpha_mode: AlphaMode,
    pub flags: SwapChainFlags,
}

description of a swapchain

Fields
width: u32

the resolution width, 0 to use the width of the CA of the target window

height: u32

the resolution height, 0 to use the height of the CA of the target window

format: DxgiFormat

the display format

stereo: Bool

whether the full-screen display mode or the back buffer is stereo

sample_desc: SampleDesc

multi-sampling scheme description

buffer_usage: Usage

surface usage and CPU access options for the back buffer. the back buffer can be used for shader input or target output.

buffer_count: u32

number of buffers in the swap chain

scaling: Scaling

scaling behavior when the back buffer is presented

swap_effect: SwapEffect

presentation model, as well as how the back buffer would be handled after calling swapchain.present().

alpha_mode: AlphaMode

transparency behavior

flags: SwapChainFlags

misc flags

Methods

impl SwapChainDesc

fn new(format: DxgiFormat) -> SwapChainDesc

create a new SwapChainDesc with default parameters