Struct redirect::device::Device

pub struct Device { /* fields omitted */ }

a 3D display adapter

Methods

impl Device

pub fn new(
adapter: Option<&Adapter>,
level: FeatureLevel
) -> Result<Device, WinError>

attempt to create a device from the given adapter and feature level. None means the default adapter would be used.

pub fn create_root_sig(
&mut self,
node_mask: u32,
desc_blob: &RootSigDescBlob
) -> Result<RootSig, WinError>

attempts to create a root signature from a description blob

pub fn create_command_queue(
&mut self,
desc: &CommandQueueDesc
) -> Result<CommandQueue, WinError>

attempts to create a command queue with given description

pub fn create_direct_command_allocator(
&mut self
) -> Result<DirectCommandAllocator, WinError>

attempts to create a direct command allocator

pub fn create_bundle_allocator(
&mut self
) -> Result<BundleCommandAllocator, WinError>

attempts to create a bundle command allocator

pub fn create_heap(&mut self, desc: &HeapDesc) -> Result<RawHeap, WinError>

attempts to create a heap

pub fn create_committed_resource(
&mut self,
heap_properties: &HeapProperties,
heap_flags: HeapFlags,
desc: &ResourceDesc,
initial_state: ResourceStates
) -> Result<RawResource, WinError>

attempts to create a committed resource

pub fn create_fence(
&mut self,
initial_value: u64,
flags: FenceFlags
) -> Result<Fence, WinError>

attempts to create a fence

pub fn get_resource_alloc_info(
&mut self,
desc: &ResourceDesc,
node_idx: u32
) -> ResourceAllocInfo

get resource allocation info from the resource description. Notice that unlike ID3D12Device::GetResourceAllocationInfo, node_idx is used instead of visibleNodeMask.

pub unsafe fn create_placed_resource(
&mut self,
heap: &mut RawHeap,
heap_offset: u64,
desc: &ResourceDesc,
initial_state: ResourceStates
) -> Result<RawResource, WinError>

attempts to create a placed resource. heap_offset must be a multiple of resource's alignment.

pub fn create_direct_command_list<'a, P: PipelineState>(
&mut self,
node_mask: u32,
alloc: &'a mut DirectCommandAllocator,
initial_state: Option<&'a P>
) -> Result<DirectCommandListRecording<'a, P>, WinError>

pub fn create_bundle<'a, P: PipelineState>(
&mut self,
node_mask: u32,
alloc: &'a mut BundleCommandAllocator,
initial_state: Option<&'a P>
) -> Result<BundleRecording<'a>, WinError>